Move composable... instructions were unclear
